Refactor Follower.
Pulled centroid-generation and publication out into a separate nodelet, leaving
Follower just responsible for listening for and responding to goal points.

In this version, the Follower nodelet subscribes to a /turtlebot_follower/goal
topic, and uses the geometry_msgs/Point message as the x, y, and z goal point
it should follow.

A separate Centroid nodelet contains the original centroid calculation logic,
and publishes its results to a topic that can be mapped to Follower's goal.

The purpose of this refactor is to allow any number of feature/object
detection-and-tracking packages to be used to generate the Follower goal, not
just the simplistic centroid calculation logic in the base implementation. A
more sophisticated person-tracking library, for example, could tell the
Turtlebot to follow only a specific person.
