+++ turtlebot_follower/cfg/Centroid.cfg
@@ -39,9 +39,11 @@ import math
 from dynamic_reconfigure.parameter_generator_catkin import *
 
 gen = ParameterGenerator()
-gen.add("goal_z", double_t, 0, "The distance away from the robot to hold the centroid.", 0.6, 0.0, 3.0)
-gen.add("x_scale", double_t, 0, "The scaling factor for translational robot speed.", 1.0, 0.0, 3.0)
-gen.add("z_scale", double_t, 0, "The scaling factor for rotational robot speed.", 5.0, 0.0, 10.0)
+gen.add("min_x", double_t, 0, "The minimum x position of the points in the box.", -0.20, -3.0,  3.0)
+gen.add("max_x", double_t, 0, "The maximum x position of the points in the box.",  0.20, -3.0, 3.0)
+gen.add("min_y", double_t, 0, "The minimum y position of the points in the box.", 0.10, -1.0, 3.0)
+gen.add("max_y", double_t, 0, "The maximum y position of the points in the box.", 0.50, -1.0, 3.0)
+gen.add("max_z", double_t, 0, "The maximum z position of the points in the box.", 0.8, 0.0, 3.0)
 
 
-exit(gen.generate(PACKAGE, "turtlebot_follower_dynamic_reconfigure", "Follower"))
+exit(gen.generate(PACKAGE, "turtlebot_follower_centroid_dynamic_reconfigure", "Centroid"))
