+++ turtlebot_follower/launch/follower.launch
@@ -2,29 +2,26 @@
   The turtlebot people (or whatever) follower nodelet.
  -->
 <launch>
-  <arg name="simulation" default="false"/>
+  <arg name="simulation"    default="false"/>
+  <arg name="generate_goal" default="true"/> <!-- Use the default goal point generation -->
+  
+  <!-- Generate Goal: Load camera and centroid goal generation nodelets -->
+  <!-- Else: Assume goal generation is loaded elsewhere -->
+  <include if="$(arg generate_goal)" file="$(find turtlebot_follower)/launch/includes/centroid.launch">
+	<arg name="simulation"  value="$(arg simulation)"/>
+  </include>
+  
+  <node pkg="nodelet" type="nodelet" name="follower_nodelet_manager" args="manager"/>
+  
   <group unless="$(arg simulation)"> <!-- Real robot -->
     <include file="$(find turtlebot_follower)/launch/includes/velocity_smoother.launch.xml">
       <arg name="nodelet_manager"  value="/mobile_base_nodelet_manager"/>
       <arg name="navigation_topic" value="/cmd_vel_mux/input/navi"/>
     </include>
-
-    <include file="$(find turtlebot_bringup)/launch/3dsensor.launch">
-      <arg name="rgb_processing"                  value="true"/>  <!-- only required if we use android client -->
-      <arg name="depth_processing"                value="true"/>
-      <arg name="depth_registered_processing"     value="false"/>
-      <arg name="depth_registration"              value="false"/>
-      <arg name="disparity_processing"            value="false"/>
-      <arg name="disparity_registered_processing" value="false"/>
-      <arg name="scan_processing"                 value="false"/>
-    </include>
   </group>
   <group if="$(arg simulation)">
-    <!-- Load nodelet manager for compatibility -->
-    <node pkg="nodelet" type="nodelet" ns="camera" name="camera_nodelet_manager" args="manager"/>
-
     <include file="$(find turtlebot_follower)/launch/includes/velocity_smoother.launch.xml">
-      <arg name="nodelet_manager"  value="camera/camera_nodelet_manager"/>
+      <arg name="nodelet_manager"  value="/follower_nodelet_manager"/>
       <arg name="navigation_topic" value="cmd_vel_mux/input/navi"/>
     </include>
   </group>
@@ -37,20 +34,12 @@
 
   <include file="$(find turtlebot_follower)/launch/includes/safety_controller.launch.xml"/>
 
-  <!--  Real robot: Load turtlebot follower into the 3d sensors nodelet manager to avoid pointcloud serializing -->
-  <!--  Simulation: Load turtlebot follower into nodelet manager for compatibility -->
   <node pkg="nodelet" type="nodelet" name="turtlebot_follower"
-        args="load turtlebot_follower/TurtlebotFollower camera/camera_nodelet_manager">
+        args="load turtlebot_follower/TurtlebotFollower follower_nodelet_manager">
     <remap from="turtlebot_follower/cmd_vel" to="follower_velocity_smoother/raw_cmd_vel"/>
-    <remap from="depth/points" to="camera/depth/points"/>
     <param name="enabled" value="true" />
     <param name="x_scale" value="7.0" />
     <param name="z_scale" value="2.0" />
-    <param name="min_x" value="-0.35" />
-    <param name="max_x" value="0.35" />
-    <param name="min_y" value="0.1" />
-    <param name="max_y" value="0.5" />
-    <param name="max_z" value="1.2" />
     <param name="goal_z" value="0.6" />
   </node>
   <!-- Launch the script which will toggle turtlebot following on and off based on a joystick button. default: on -->
